[Bugfix] Fixed the `DATA_VALID` assertion on chained writes.
There was an issue with the current structure related to
when `DATA_VALID` would assert. This happened when we chained
two long writes together (future testbench coming). The `DATA_VALID`
would be delayed by one JTAG state transition and hence would
not allow the user time to read the value before an overflow
occurred.

I've refactored the read-back section into separate processes
for control and shift registers. This makes it easier to
capture the correct logic for the signls in question.