Added initial implementation of a Fifo VC
This adds a new verification component for use with
testing systems that need to use the Xilinx FIFO IP. The
idea is that this verification component can be used in place
of that generated IP.

This is setup to use the `OSVVM` libraries. See here:

https://github.com/OSVVM/OsvvmLibraries

I have not setup a git submodule for this yet. I'm going to
do that in a subsequent step.