[Refactor] Reorganized the test suite
This should make it easier to add more test files in the future.

Note - that the name of the file, the name of the config, and
the string passed to the `SetTestName` function all need to match.
So I've decided to make the arch, configuration, and filename
all use the same string.

I've also added a new testcase that evaluates the blocking
Send operation in the FifoVC component