[Refactor] Moved tools and JTAG dependencies into a library.
I was having an issue where I was importing the `src/JTAG.vhd` file
in both the Jtag Device VC library and in a new TAPController
testbench library. Attempting interop between these two was impossible
because the JTAG definitions in the JtagDevVC library and the
TAPController library were not the same.

By moving them into a shared common library I can work around this
issue.

Note - careful with including `*.pro` files that create libraries.
There is no scoping. So the `library` definition in the included
`*.pro` file will stick until you reset it.