Added initial OSVVM structure for TAPController unit tests
This adds in a new test structure for the TAPController. I'm
using the OSVVM framework to coordinate the test sequences.

Note that I've only brought one test in. I'm going to bring
in the other tests and cleanup the repo in a separate commit.

The current `Makefile` based build is broken at this point.